requirements.txt
flask
connexion
requests
werkzeug
PyJWT
